Added styling for all recipes
All recipes use the same style sheet for consistency.
The heading is written in orange-ish color and the text is aligned at
the center of the page.
The background has a very dark gray color.
The image has been given a width of 300px and height is set to auto.
The sub-headings have a dark green color.
The text in paragraphs and lists have all been given a color of white.